[fix] Extract PROGRAMS catalog and split OnboardingFlow into per-step components
Addresses two /review findings on PR #160:

(1) PROGRAMS catalog (~105 lines, 13 entries with prose descriptions
    and metadata) was embedded inline in the client component. The
    moment Step 4 wires to the API, the same catalog will need to live
    server-side too, and packages/core already references "5/3/1" in
    slotMaps.ts. Extracted to ./programs.ts as a sibling data module
    so the future migration to packages/core is a one-line import
    change.

(2) OnboardingFlow.tsx was a 322-line monolith owning the 4-step state
    machine, lift form, 1RM calculator, program filter/detail toggle,
    and success banner via {step === N && (...)} blocks. Step 4 will
    roughly double in size when API wiring lands. Split into:
      - lib.ts                 — shared types (DiscoveryMethod, LiftKey,
                                  LiftEntry), LIFT_LABELS, brzycki1RM
      - programs.ts            — Experience, Program, PROGRAMS
      - steps/StepMethod.tsx   — method picker (METHOD_OPTIONS local)
      - steps/StepLifts.tsx    — lift entry form
      - steps/StepConfirm.tsx  — computed maxes display
      - steps/StepProgram.tsx  — program filter, detail, confirm
      - OnboardingFlow.tsx     — state orchestrator (~120 lines)

OnboardingFlow now owns state and delegates rendering to step
components via props. Per-step files become independently testable
and the API-wiring PR will be a localized diff in StepProgram.tsx
rather than another edit to a 500-line file.
